requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.8
python-dotenv==1.0.0
phonenumbers==8.13.27
django-filter==23.5
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
twilio==8.9.1
retry==0.9.2
